addressing requested changes
