+++ crates/transaction-pool/src/pool/mod.rs
@@ -111,7 +111,7 @@ pub use parked::{BasefeeOrd, ParkedOrd, ParkedPool};
 pub use pending::PendingPool;
 
 mod best;
-mod best_helper;
+mod best_filter;
 mod blob;
 pub use blob::{blob_tx_priority, fee_delta};
 mod parked;
